% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacks.R
\name{add_reveal_row}
\alias{add_reveal_row}
\title{Add a duplicate final row}
\usage{
add_reveal_row(df, reveal_index)
}
\arguments{
\item{df}{the dataframe}

\item{reveal_index}{the name of the variable being used}
}
\description{
Takes the final row (ie with the highest value of `reveal_index`) of the
dataframe `df`, adds +1 to `reveal_index` and then binds that duplicate
final row onto the bottom of `df`.
}
