% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacks.R
\name{add_zero_state}
\alias{add_zero_state}
\title{Duplicate dataset with zero values for one variable}
\usage{
add_zero_state(df, value)
}
\arguments{
\item{df}{the dataframe}

\item{value}{the variable that will be reset as zero in the copy dataset.}
}
\description{
This function duplicates a dataset, using a variable `state` to
distinguish the two copies:
the original version has `state=1`, the copy has `state=0`.
}
\details{
The copy also sets `value = 0` throughout.
Useful in conjunction with `transition_state()`.
Eg for a bar plot, state = 1, the actual values, where an animation ends.
state = 0, all plotted values set to 0, where an animation starts.
}
