% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_ons.R
\name{bib_ons}
\alias{bib_ons}
\title{BiBTeX ref for an ONS publication}
\usage{
bib_ons(tag = "tag", url = clipr::read_clip())
}
\arguments{
\item{tag}{The citation label to be used to refer to the
publication in (e.g.) .qmds}

\item{url}{the webpage for the publication. Default is to take
it from the clipboard.}
}
\description{
Produce a BibTeX citation for an ONS publication.
}
\examples{

the_url <- "https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/bulletins/crimeinenglandandwales/yearendingjune2023"
ons_ref(tag = "the_label", url = the_url)
}
