% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{checkscel}
\alias{checkscel}
\title{open a spreadsheet}
\usage{
checkscel(path, app = c("Excel", "Numbers"))
}
\arguments{
\item{path}{where it is}

\item{app}{how to open it: Excel or Numbers}
}
\description{
open a spreadsheet
}
