% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{chexcel}
\alias{chexcel}
\title{open a spreadsheet}
\usage{
chexcel(path, app = c("Excel", "Numbers"))
}
\arguments{
\item{path}{where it is}

\item{app}{how to open it: Excel or Numbers}
}
\description{
Tested on mac only
}
