% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_string.R
\name{colour_string}
\alias{colour_string}
\title{Prepare multicolour string for ggtext}
\usage{
colour_string(original, to_colour = NULL)
}
\arguments{
\item{original}{the string of text (e.g. a plot title or annotation)}

\item{to_colour}{a named vector of colours. The names of each colour are
the words in the text that will be drawn that colour.}
}
\value{
the string with html tags to add colour to the specified words
}
\description{
Writes html tags around phrases in a string to show those
phrases in a different colour
}
