% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_functions.R
\name{cut_at}
\alias{cut_at}
\title{Cut a vector at breakpoints defined by quantiles}
\usage{
cut_at(x, at, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{at}{The quantiles at which to break x
(defined either in the range 0-100 or 0-1).}

\item{...}{other arguments passed to \code{\link{cut}}}
}
\value{
An ordered factor, dividing x at the quantiles given by \code{at}. Labels are of the form (e.g.) "0-25\%", "25-33\%"...
}
\description{
Cut a vector at breakpoints defined by quantiles
}
\examples{
cut_at(100:1, c(25, 75))
}
\seealso{
\code{\link{cut_into_n}}
}
