% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_functions.R
\name{cut_into_n}
\alias{cut_into_n}
\title{Cut a vector into n equal parts}
\usage{
cut_into_n(x, n, labels = paste0(1:n, "of", n), ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of equal sized parts.}

\item{labels}{Labels for the factor passed to \code{\link{cut}}.
Default value is  "1ofn", "2ofn", ..."nofn"}

\item{...}{other arguments passed to \code{\link{cut}}}
}
\value{
An ordered factor, dividing x in to n-ths, with default values
"1ofn" < "2ofn", < ... < "nofn"
}
\description{
Cut a vector into n equal parts
}
\examples{
cut_into_n(100:1, 3)
}
\seealso{
\code{\link{cut_at}}
}
