% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{dbdir}
\alias{dbdir}
\title{Open a directory in Finder}
\usage{
dbdir(dir = config$db_proj)
}
\arguments{
\item{dir}{The name of the directory. Default is the Dropbox project dir.}
}
\description{
Does a bit of work to allow for BBC Dropbox folder names
}
