% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacks.R
\name{ddate_breaks}
\alias{ddate_breaks}
\title{Work out breaks for an axis of decimal dates}
\usage{
ddate_breaks(x)
}
\arguments{
\item{x}{the range over which the dates should be calculated}
}
\description{
gganimate can't zoom across dates.
So, if zooming, a chart will need to be based on decimal dates
(e.g. 2022.01, 2022.7).
}
\details{
That causes problems for picking sensible dates for breaks
(e.g. why prefer Jan 1 over Jan 6 when they're decimals?)
This function calculates those breaks.

The break width changes depends on the width of the data. Rather than using
R's pretty algorithm, it manually specifies whether it's month, year etc.
}
