% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{download.new.file}
\alias{download.new.file}
\title{Download a new version of a file}
\usage{
download.new.file(url, fn)
}
\arguments{
\item{url}{the file to be downloaded}

\item{fn}{the filename where it will go. Won't download unless the online file has been updated more
recently than this file has been modified.}
}
\description{
Uses `curl -z` to check that the online file
has been updated more recently than the timestamp on the
destination file
}
