% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{calculate the centre and width/height of a data range.}
\usage{
make_tiles(
  df,
  x_var,
  y_var,
  xstretch = 1.2,
  ystretch = 2,
  add_animation_vars = TRUE
)
}
\arguments{
\item{df}{the data frame}

\item{x_var, y_var}{the variables to be summarised}

\item{xstretch, ystretch}{how much extra to expand the width and height
of the tiles. Set to 1 for no expansion. NB expanding the data
range might expand  the range of your chart.}

\item{add_animation_vars}{add a selection of variables to aid tweening?}
}
\description{
Used to help make tiles.
}
\examples{
data <- data.frame(x1 = runif(10), y2=runif(10))
tiles <- make_tiles(data, x1, y2, ystretch=1, add=FALSE)

library(ggplot2)
ggplot(data, aes(x=x1, y=y2)) +
 geom_tile(data=tiles, aes(width=width, height=height),
   colour="red", fill=NA) + 
 geom_point()

make_tiles(data, x1, y2)
}
