% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_date.R
\name{month_to_date}
\alias{month_to_date}
\title{Convert a string containing month-and-year to a lubridate date}
\usage{
month_to_date(months, when = c("beginning", "middle", "end"), format = "dmy")
}
\arguments{
\item{months}{A character vector in a month/year format that lubridate might recognise}

\item{when}{A flag determining whether you want the get the 1st ("beginning"),
15th ("middle") or last ("end") day of the month}

\item{format}{A string for the lubridate function that you want to turn your character
string into a date (should always start with "d")}
}
\value{
a vector of dates
}
\description{
Convert a string containing month-and-year to a lubridate date
}
\examples{
month_to_date("jun2017")
month_to_date("jun2017", "e")
month_to_date("2017/jun", format="dym")
}
