% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{plot_window}
\alias{plot_window}
\title{A plot preview window that shows if your title fits}
\usage{
plot_window(
  form = c("online", "tv", "portrait"),
  width = NULL,
  height = NULL,
  title = NULL
)
}
\arguments{
\item{form}{template for the sizing of the output}

\item{width, height, title}{finer control over sizing/title for plot window.
In inches. To convert from pixel size, divide by 72.}
}
\description{
Plotting windows that match default dimensions for BBC outputs
}
\examples{

plot_window("online")
plot_window("tv")
plot_window("online", height=800/72)

}
