% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{pptx}
\alias{pptx}
\title{Arrange alignment and save BBC ggplot chart as powerpoint}
\usage{
pptx(
  plot_name,
  source_name = NULL,
  save_filepath = "~/Downloads/tmp.pptx",
  pptx_template = file.path(system.file("data", package = "rcutils"), "layouts.pptx"),
  ...
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created that you want to format and save}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your side}

\item{save_filepath}{Exact filepath that you want the plot to be saved to}

\item{pptx_template}{The location of the powerpoint template that
officer will use to build your powerpoint slides. Best to stick with
the default}
}
\value{
the filename of the new powerpoint slide
}
\description{
Running this function will save your plot into powerpoint.
You need the packages rvg, officer and bbplot2
}
