% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{quantiles}
\alias{quantiles}
\title{Old function for splitting continuous vectors into chunks.}
\usage{
quantiles(x, n, include.lowest = T, ...)
}
\description{
Replaced by \code{\link{cut_into_n}} and \code{\link{cut_at}}
}
