% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reith_report_style.R
\name{reith_report_style}
\alias{reith_report_style}
\title{Reith-ish styling for figures in .Rmd/.qmd reports}
\usage{
reith_report_style(...)
}
\arguments{
\item{...}{arguments being passed to [ggplot2::theme()]}
}
\description{
Reith-ish styling for figures in .Rmd/.qmd reports
}
