% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{safe_fn}
\alias{safe_fn}
\title{fix the spaces and brackets in Dropbox (BBC) and Visual Journalism
so that they don't break the command line arguments of system2()}
\usage{
safe_fn(path)
}
\arguments{
\item{path}{the file path to be made safe}
}
\description{
fix the spaces and brackets in Dropbox (BBC) and Visual Journalism
so that they don't break the command line arguments of system2()
}
