% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{shrink}
\alias{shrink}
\title{A smaller range}
\usage{
shrink(rng, x)
}
\arguments{
\item{rng}{The original range}

\item{x}{How much to shrink it by}
}
\value{
A new, smaller range
}
\description{
A smaller range
}
