% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_newsroom2.R
\name{skew}
\alias{skew}
\title{Adjust a graphic for fish-eye cameras}
\usage{
skew(infile, outfile, across = 100, squeeze = 200, split = 0.5, bg = "#f5f5f5")
}
\arguments{
\item{infile}{the png you want to squeeze}

\item{outfile}{where you want to save the output}

\item{across}{how many pixels you want to shove it to the right}

\item{squeeze}{how many pixels you want to squeeze the RHS in by. Half
will be taken off the top and half off the bottom.}

\item{split}{how much of the `squeeze` pixels are taken at the top. Default
makes for a 50:50 split. Defined in the range 0-1.}

\item{bg}{the background colour you want to use to fill in any blank space
created.}
}
\description{
Squeezes the right-hand-side of an image to make it look
normal on telly when filmed from a fish-eye camera.
}
\details{
You need to have Magick installed for this to work
(the software, not the R package)
}
\examples{
\dontrun{
## shift everything 100 pixels left, take 100 off the top
## and 100 off the bottom.
skew("file1.png", "file2.png")
## shift everything 100 pixels left, take 150 off the top
## and 50 off the bottom.
skew("file1.png", "file2.png", split = .75)
## shift everything 200 pixels left, take 75 off the top and 75
## off the bottom.
skew("file1.png", "file2.png", across = 200, squeeze = 150)
}
}
