% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retag.R
\name{sretag}
\alias{sretag}
\title{Strip a live post url to minimum details}
\usage{
sretag(url, new_path = "/news/reality_check", goto = TRUE)
}
\arguments{
\item{url}{the live page post url}

\item{new_path}{a new live page or topic page to which 
you want to point the reader. Set to `NULL` if 
you want to leave as is.}

\item{goto}{Do you want to open that new page in a browser (Mac only) or
get the new url back as a message}
}
\description{
Strip a live post url to minimum details
}
\examples{

url <- "https://www.bbc.co.uk/news/live/business-66861406?ns_mchannel=social&ns_source=twitter&ns_campaign=bbc_live&ns_linkname=650aa0c63fc96e28602bd790\%26The\%20surprising\%20reason\%20food\%20inflation\%20is\%20slightly\%20overstated...\%262023-09-20T08\%3A04\%3A03.589Z&ns_fee=0&pinned_post_locator=urn:asset:4fd2ba06-a0d5-4784-8825-2f5b0024f806&pinned_post_asset_id=650aa0c63fc96e28602bd790&pinned_post_type=share"
retag(url, goto = FALSE)
retag(url, new_path = NULL, goto = FALSE)
retag(url)
retag(url, new_path = NULL)
}
