% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacks.R
\name{thin}
\alias{thin}
\title{Thin a vector}
\usage{
thin(x, jump_size = NULL, n = NULL, neg = FALSE)
}
\arguments{
\item{x}{the vector to be thinned}

\item{jump_size}{pick every `jump_size`th element}

\item{n}{pick (roughly) length elements, ie set `jumps = length(x) / n.`}

\item{neg}{Return the elements picked (default behaviour) or
the complement (ie all the other elements).}
}
\description{
Take every (say) fifth element from a vector (like a list of
filenames, keeping the first and final ones.
}
\details{
If you've got 200 frames for a video and want to thin it to around 40,
then use this.
}
\examples{
thin(1:19, jump = 5)
thin(1:20, n = 6)
}
