% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{tidyxls_sheet}
\alias{tidyxls_sheet}
\title{Write an xls sheet to xlsx so you can open it with tidyxl}
\usage{
tidyxls_sheet(fn, sheet)
}
\arguments{
\item{fn}{the filename of a .xls spreadsheet}

\item{sheet}{the worksheet to be xlsx_cells'ed.}
}
\description{
Write an xls sheet to xlsx so you can open it with tidyxl
}
