% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok_preview.R
\name{tiktok_preview}
\alias{tiktok_preview}
\title{Check whether a layout fits in tik-tok safe spaces}
\usage{
tiktok_preview(plot)
}
\arguments{
\item{plot}{a ggplot object, fully styled and ready to go.}
}
\value{
(Invisibly) the pathname of the generated png.
}
\description{
Takes your plot, saves it out at 1080x1920px,
overlays a tiktok view of buttons etc. on top of that
and then previews the resultant image.
}
\examples{

library(ggplot2)

plot <- ggplot(uk_gdp, aes(x = date, y = value)) +
  geom_line(colour = tvcols["news"], lineend = "round") +
  labs(
    title = "Quarterly UK GDP",
    caption = "Source: Office for National Statistics"
  ) +
  portrait_style(
    plot_margins = c(252, 240, 400, 80),
    bottom = 250
  ) +
  scale_y_continuous(
    label = scales::label_dollar(
      prefix = "£",
      scale_cut = c(" " = 0, "bn" = 1e3, "tn" = 1e6)
    )
  )

\dontrun{

tiktok_preview(plot)

tiktok_preview(plot + portrait_style(plot_margins = c(150, 20, 20, 20)))
}
}
