% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{views}
\alias{views}
\title{Look at a dataframe using a spreadsheet package. Mac only.}
\usage{
views(df, app = c("Numbers", "Excel"))
}
\arguments{
\item{df}{the data frame}

\item{app}{the program used to show it. Defaults to Apple's Numbers. Excel also an option.}
}
\value{
nothing, just opens a spreadsheet showing the dataframe.
}
\description{
You can't edit the dataframe - it just writes the dataframe to a tempfile which it then opens.
}
\examples{

demo <- data.frame(alpha=letters, num=rnorm(26))
views(demo)
views(demo, "Excel")
views(demo, "E")
}
