% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_helpers.R
\name{cut_into_n}
\alias{cut_into_n}
\title{Cut a vector into N equal parts}
\usage{
cut_into_n(x, n, include.lowest = T, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of equal sized parts.}

\item{include.lowest}{Passed to \code{\link{cut}}. The default value used here (T)
overwrites \code{cut}'s usual behaviour
of excluding the lowest value in the dataset from any group.}

\item{...}{other arguments passed to \code{\link{cut}}}
}
\value{
a character vector, dividing x in to Nths, with values  "1 of N", "2 of N", ... "N of N"
}
\description{
Cut a vector into N equal parts
}
