% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd_structure.R
\name{oecd_structure}
\alias{oecd_structure}
\title{Get the variable names and code/decode pairs for an OECD dataset}
\usage{
oecd_structure(dataset)
}
\arguments{
\item{dataset}{the name of the dataset on the OECD data system. For a list of
OECD datasets, use OECD::get_datasets()}
}
\description{
Get the variable names and code/decode pairs for an OECD dataset
}
\examples{
oecd_structure("NAAG")
oecd_structure("SNA_TABLE8A")
}
